Remove upload page
I am currently unsure how I want to use this page. Maybe a upload from device, but this does not seem useful on the moble. I real want to integrate a social system like goodreads, with messaging, and the ablity to share books. Unfornicalty the database side was not desided for this. I would like to rebuild the whole thing in order to better serve this purpose